Input files names taken from argv
If no parameters are passed, the default name is used. This is how the software was working.
If one or more file names are provided, they are processed in the same order.
Some options can be passed (TODO). If only options are passed, the behavior is the same as no parameters.